ai-service/dog-detection/requirements.txt
tensorflow>=2.5.0
Pillow
numpy
